src/main/java/kitchenpos/menu/application/MenuGroupService.java
package kitchenpos.menu.application;

import java.util.List;

import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import kitchenpos.menu.domain.MenuGroup;
import kitchenpos.menu.domain.MenuGroupRepository;
import kitchenpos.menu.dto.MenuGroupRequest;

@Service
@Transactional(readOnly = true)
public class MenuGroupService {
	private final MenuGroupRepository menuGroupRepository;

	public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
		this.menuGroupRepository = menuGroupRepository;
	}

	@Transactional
	public MenuGroup create(final MenuGroupRequest menuGroup) {
		return menuGroupRepository.save(menuGroup.toEntity());
	}

	public List<MenuGroup> list() {
		return menuGroupRepository.findAll();
	}
}
